+++ WindowsFormsApplication11/Menu_Item_Price.cs
@@ -21,7 +21,7 @@ namespace WindowsFormsApplication11
         }
     
         public int Menu_Price_ID { get; set; }
-        public decimal Menu_Price { get; set; }
+        public double Menu_Price { get; set; }
         public System.DateTime Date { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
